update the blade callback file
